Move LiveGame|Player into live.ts model file
