+++ src/components/lineup-game-setup.ts
@@ -11,7 +11,8 @@ import { repeat } from 'lit/directives/repeat.js';
 import { navigate } from '../actions/app';
 import { connectStore } from '../middleware/connect-mixin';
 import { FormationBuilder, FormationMetadata, formatPosition, Position } from '../models/formation';
-import { GameDetail, LivePlayer, SetupStatus, SetupSteps, SetupTask } from '../models/game';
+import { GameDetail, SetupStatus, SetupSteps, SetupTask } from '../models/game.js';
+import { LivePlayer } from '../models/live.js';
 import { getGameStore } from '../slices/game-store';
 import { gameSetupCompletedCreator } from '../slices/game/game-slice.js';
 import { getLiveStore } from '../slices/live-store';
