+++ test/unit/helpers/test-live-game-data.ts
@@ -1,7 +1,7 @@
-import { LivePlayer, LiveGame, GameStatus } from '@app/models/game';
-import { LiveGames } from '@app/models/live.js';
-import { Player, PlayerStatus } from '@app/models/player';
-import { STORED_GAME_ID } from './test_data';
+import { GameStatus } from '@app/models/game.js';
+import { LiveGame, LiveGames, LivePlayer } from '@app/models/live.js';
+import { Player, PlayerStatus } from '@app/models/player.js';
+import { STORED_GAME_ID } from './test_data.js';
 
 export function buildLiveGames(games: LiveGame[]): LiveGames {
   return games.reduce((obj, game) => {
