+++ test/unit/slices/live/shift-slice.test.ts
@@ -1,5 +1,5 @@
 import { Duration } from '@app/models/clock.js';
-import { LivePlayer } from '@app/models/game.js';
+import { LivePlayer } from '@app/models/live.js';
 import { PlayerStatus } from '@app/models/player.js';
 import { PlayerTimeTrackerMap } from '@app/models/shift.js';
 import { endPeriod, startPeriod } from '@app/slices/live/clock-reducer-logic.js';
